Refine admin flows and rename panels
