+++ admin/public/config.js.template
@@ -1,3 +1,4 @@
 window.BACKEND_URL = "${BACKEND_URL}";
 window.TWITCH_CLIENT_ID = "${TWITCH_CLIENT_ID}";
 window.TWITCH_SCOPES = "${TWITCH_SCOPES}";
+window.BOT_APP_SCOPES = "${BOT_APP_SCOPES}";
