+++ example.env
@@ -29,9 +29,16 @@ TWITCH_REDIRECT_URI=http://localhost:8000/auth/callback
 # Scopes must include channel:bot for EventSub
 TWITCH_SCOPES="chat:read chat:edit channel:bot"
 
+# Scopes requested when the Admin panel generates a bot app token
+BOT_APP_SCOPES="user:read:chat user:write:chat user:bot"
+
 # EventSub configuration
 TWITCH_EVENTSUB_CALLBACK=http://backend:7070/eventsub/callback
 TWITCH_EVENTSUB_SECRET=change-me
 
 # Port for exposing the web interface (optional)
 WEB_PORT=7000
+
+# Optional port overrides for management panels
+QUEUE_MANAGER_PORT=7100
+ADMIN_PORT=7171
