+++ queue_manager/nginx.conf
