+++ queue_manager/public/config.js.template
