+++ queue_manager/public/index.html
@@ -1,0 +1,70 @@
+<!doctype html>
+<html>
+<head>
+  <meta charset="utf-8"/>
+  <meta name="viewport" content="width=device-width, initial-scale=1"/>
+  <title>ALPEN.BOT SONGQ</title>
+  <link rel="stylesheet" href="style.css"/>
+</head>
+<body>
+<div id="login-status" class="status-bar" style="display:none">
+  <div class="status-user">
+    <div id="login-avatar" class="status-avatar" aria-hidden="true"></div>
+    <div class="status-info">
+      <span id="login-name" class="status-name"></span>
+      <span id="login-channel" class="status-channel"></span>
+    </div>
+  </div>
+  <div class="status-actions">
+    <button id="logout-btn" class="status-action">log out</button>
+    <button id="logout-perm-btn" class="status-action danger">forget &amp; log out</button>
+  </div>
+</div>
+<div id="landing" class="landing">
+  <h1>ALPEN.BOT SONGQ</h1>
+  <button id="login-btn">log in with Twitch</button>
+  <div id="channel-list" class="channel-list"></div>
+</div>
+
+<div id="app" style="display:none">
+  <div class="wrap">
+    <div class="header">
+      <div class="brand">Queue Manager</div>
+      <span id="ch-badge" class="badge">channel: ?</span>
+      <button id="reg-btn" style="display:none"></button>
+    </div>
+
+    <div class="grid">
+      <div class="card">
+        <div class="tabs">
+          <button id="tab-queue" class="tab active">Queue</button>
+          <button id="tab-users" class="tab">Users</button>
+          <button id="tab-settings" class="tab">Settings</button>
+        </div>
+
+        <div id="queue-view">
+          <h2>Active Queue</h2>
+          <div id="queue" class="queue"></div>
+          <div class="footer">
+            <button id="archive-btn">Archive Stream</button>
+            <button id="mute-btn">Toggle Requests</button>
+          </div>
+        </div>
+
+        <div id="users-view" style="display:none">
+          <h2>Users</h2>
+          <div id="users" class="queue"></div>
+        </div>
+
+        <div id="settings-view" style="display:none">
+          <h2>Settings</h2>
+          <div id="settings"></div>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
+  <script src="config.js"></script>
+  <script src="queue_manager.js"></script>
+</body>
+</html>
