--- stats/public/index.html
@@ -1,94 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset="utf-8"/>
-  <meta name="viewport" content="width=device-width, initial-scale=1"/>
-  <title>ALPEN.BOT SONGQ — Host Stats</title>
-  <link rel="stylesheet" href="style.css"/>
-</head>
-<body>
-  <div class="wrap">
-    <header class="header">
-      <div class="brand">Hoster Statistics</div>
-      <span id="status" class="badge">status: loading…</span>
-    </header>
-
-    <section class="card">
-      <h2>Registered Channels</h2>
-      <div id="channel-list" class="channel-list"></div>
-    </section>
-
-    <section class="card">
-      <h2>Channel Overview</h2>
-      <div id="channel-tree" class="tree"></div>
-    </section>
-
-    <section class="card bot-card" id="bot-card">
-      <div class="card-head">
-        <h2>Bot Control</h2>
-        <div class="bot-actions">
-          <button type="button" id="bot-login-btn">Sign in</button>
-          <button type="button" id="bot-logout-btn" hidden>Sign out</button>
-          <button type="button" id="bot-authorize-btn" disabled>Authorize Bot</button>
-        </div>
-      </div>
-      <div id="bot-guard" class="bot-guard">
-        <p>Sign in with Twitch to manage the bot configuration.</p>
-      </div>
-      <div id="bot-panel" class="bot-panel" hidden>
-        <div class="bot-meta">
-          <div class="bot-meta-block">
-            <span class="label">Signed in as</span>
-            <span id="bot-owner" class="value">—</span>
-          </div>
-          <div class="bot-meta-block">
-            <span class="label">Authorized bot</span>
-            <span id="bot-account" class="value">Not connected</span>
-          </div>
-          <div class="bot-meta-block">
-            <span class="label">Token expires</span>
-            <span id="bot-expiry" class="value">—</span>
-          </div>
-        </div>
-        <div id="bot-alert" class="bot-alert" hidden></div>
-        <div class="bot-settings">
-          <label class="field">
-            <span>Bot account name</span>
-            <input type="text" id="bot-display-name" placeholder="Display name" autocomplete="off"/>
-          </label>
-          <label class="toggle">
-            <input type="checkbox" id="bot-enabled"/>
-            <span>Enable bot</span>
-          </label>
-          <div class="scope-picker">
-            <div class="scope-head">
-              <span>Scopes</span>
-              <div class="scope-actions">
-                <button type="button" id="bot-scope-reset">Reset</button>
-              </div>
-            </div>
-            <div id="bot-scope-list" class="scope-list"></div>
-            <div class="scope-add">
-              <input type="text" id="bot-scope-custom" placeholder="Add scope" autocomplete="off"/>
-              <button type="button" id="bot-scope-add">Add</button>
-            </div>
-          </div>
-        </div>
-        <div class="bot-console">
-          <div class="bot-console-head">
-            <h3>Console</h3>
-            <div class="console-actions">
-              <span id="bot-console-status" class="badge">status: idle</span>
-              <button type="button" id="bot-console-clear">Clear</button>
-            </div>
-          </div>
-          <div id="bot-console-log" class="bot-console-log"></div>
-        </div>
-      </div>
-    </section>
-  </div>
-
-  <script src="config.js"></script>
-  <script src="stats.js"></script>
-</body>
-</html>
